Added a reset breifs to default button --2

+++ app/Models/User.php
@@ -1941,20 +1941,24 @@ class User extends Authenticatable
 
     public function resetQuestions(){
 
-        $agentQuestionnaires = AgentQuestionnaire::where(['agent_id' => $this->id])->get();
-        foreach($agentQuestionnaires as $agentQuestionnaire){
-            $agentQuestionnaireElements = AgentQuestionnaireElement::where(['agent_questionnaire_id' => $agentQuestionnaire->id])->get();
+        $platformManagerId = config('constants.admins')['1platformagent']['user_id'];
 
-            if (count($agentQuestionnaireElements)) {
-                foreach($agentQuestionnaireElements as $agentQuestionnaireElement){
-                    $agentQuestionnaireElement->delete();
+        if ($this->id != $platformManagerId) {
+            $agentQuestionnaires = AgentQuestionnaire::where(['agent_id' => $this->id])->get();
+            foreach($agentQuestionnaires as $agentQuestionnaire){
+                $agentQuestionnaireElements = AgentQuestionnaireElement::where(['agent_questionnaire_id' => $agentQuestionnaire->id])->get();
+
+                if (count($agentQuestionnaireElements)) {
+                    foreach($agentQuestionnaireElements as $agentQuestionnaireElement){
+                        $agentQuestionnaireElement->delete();
+                    }
                 }
+
+                $agentQuestionnaire->delete();
             }
 
-            $agentQuestionnaire->delete();
+            $this->createDefaultQuestions();
         }
-
-        $this->createDefaultQuestions();
     }
 }
